+++ migrations/20160712130604_development.js
@@ -6,7 +6,7 @@ exports.up = function(knex, Promise) {
         .createTableIfNotExists('account_invitations', function(table) {
           table.increments('id').primary();
           table.uuid('uuid').unique().defaultTo(knex.raw('uuid_generate_v4()'));
-          table.integer('account_id').references('id').inTable('accounts').onDelete('CASCADE');
+          table.bigint('account_id').references('id').inTable('accounts').onDelete('CASCADE');
           table.string('email');
           table.specificType('invited_role_ids','int[]');
           table.dateTime('date_invited').defaultTo(knex.fn.now());
